requirements.txt
asgiref==3.2.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
circleci==1.2.2
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.9
djangorestframework==3.11.0
djongo==1.3.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
npm==0.1.1
optional-django==0.1.0
pymongo==3.10.1
python-dateutil==1.5
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
